requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
textblob==0.17.1
python-multipart==0.0.6
yfinance==0.2.28
httpx==0.25.2
openai==1.12.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
tweepy==4.14.0
snscrape==0.7.0.20230622
